haze: refactor constant use for cleaner separation
