+++ troposphere/haze/include/haze/ptp_responder.hpp
@@ -19,6 +19,7 @@
 #include <haze/async_usb_server.hpp>
 #include <haze/ptp_object_heap.hpp>
 #include <haze/ptp_object_database.hpp>
+#include <haze/ptp_responder_types.hpp>
 
 namespace haze {
 
@@ -30,12 +31,13 @@ namespace haze {
             FileSystemProxy m_fs;
             PtpUsbBulkContainer m_request_header;
             PtpObjectHeap *m_object_heap;
+            PtpBuffers* m_buffers;
             u32 m_send_object_id;
             bool m_session_open;
 
             PtpObjectDatabase m_object_database;
         public:
-            constexpr explicit PtpResponder() : m_usb_server(), m_fs(), m_request_header(), m_object_heap(), m_send_object_id(), m_session_open(), m_object_database() { /* ... */ }
+            constexpr explicit PtpResponder() : m_usb_server(), m_fs(), m_request_header(), m_object_heap(), m_buffers(), m_send_object_id(), m_session_open(), m_object_database() { /* ... */ }
 
             Result Initialize(EventReactor *reactor, PtpObjectHeap *object_heap);
             void Finalize();
@@ -48,9 +50,13 @@ namespace haze {
             Result HandleCommandRequest(PtpDataParser &dp);
             void ForceCloseSession();
 
-            template <typename Data>
-            Result WriteResponse(PtpResponseCode code, Data &&data);
+            Result WriteResponse(PtpResponseCode code, const void* data, size_t size);
             Result WriteResponse(PtpResponseCode code);
+
+            template <typename Data> requires (util::is_pod<Data>::value)
+            Result WriteResponse(PtpResponseCode code, const Data &data) {
+                R_RETURN(this->WriteResponse(code, std::addressof(data), sizeof(data)));
+            }
 
             /* PTP operations. */
             Result GetDeviceInfo(PtpDataParser &dp);
